[FIX] web_gantt: non deterministic tests
the helper setScale is no longer async and triggers a change in the gantt
model and a gantt renderer rendering after some milliseconds (see debounce
in GanttRendererControls). If we do not wait that rendering to select
a new time range via the date pickers, those date pickers can close too
early and make the test helper crash.
To fix that problem, we use the helper ganttControlsChanges that run all
timers and wait for a suitable number of animation frames.

Runbot issue: 67941

closes odoo/enterprise#64598

X-original-commit: e447315d6266d65202c869c4c7bbc2297d0bc4e9
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>