[FIX] account_reports: closing entry refresh move
Before this commit, when posting a manually modified closing entry, the entire
move would be refreshed, and all user-modified information would be erased.
This occurred because, when accessing the action_periodic_vat_entries, the
current tax closing move was overwritten.

This commit changes this behavior by adding a flag to the method, indicating
whether it is called from the _close_tax_period function. If so, it only updates
moves that are different from the tax closing of the current company.

For the depending moves, instead of posting them we now display a custom
component with a link to the depending moves.

Additionally, when selecting only one company within a tax unit, it sometimes
happened that, when closing the main company, the dependent closings to post
(depending_closings_to_post) were empty because the moves were not generated
yet for the other companies.

Also removing the banner tax_closing_show_multi_closing_warning (will remove the
 field in another pr in master)

closes odoo/enterprise#67588

Task: 3987807
X-original-commit: a421702db0506b5f1981f09183eccbaa059ca379
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Maximilien La Barre (malb) <malb@odoo.com>