[FIX] documents_spreadsheet_crm: fix dependencies
The module depends on documents_spreadsheet and crm in the manifest
but contains referneces for fields defined in crm_enterprise and relies
on the autoinstall of the enterprise module. The dependency should be
explicit because crm_enterprised can be uninstalled by users and the
documents would still be there with missing fields.

closes odoo/enterprise#65917

X-original-commit: f200be9402dbac3a80e15bcc5143362a4a7a6262
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>
Signed-off-by: Jinane Maksoud (maji) <maji@odoo.com>