[IMP] mrp_workorder: shop floor ui/ux improvements
This commit adds minor ui/ux changes to the "Shop Floor" app:
When adding a component or a by-product to an MO from the shop floor, the product
catalog is opened instead of the existing wizard, similar to adding
them from the MO form. The components that are added from the shop floor
are flagged to be manually consumed.

Task-3962134

Part-of: odoo/enterprise#64600
Related: odoo/odoo#169258
Related: odoo/upgrade#6186
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>