[IMP] spreadsheet_edition: change locale globe warning color
there's currently a small globe on the top-right corner displayed when the
spreadsheet locale is different than the user's locale. The goal is to draw
attention of the user that he needs to write numbers and dates accordingly.

The globe is currently red, which might be scary. The user might think
something is wrong. This task changes the globe color to a more neutral "info"
blue.

closes odoo/enterprise#64579

Task: 3989485
Signed-off-by: Pierre Rousseau (pro) <pro@odoo.com>