[IMP] various: change the date range in gantt view.
various : hr_attendance_gannt, hr_work_entry_contract_enterprise

Following the gantt view rework, you see directly 3 periods
of time on your selection, 3 months, 3 days, 3 years, ... 
But, in attendance and payroll apps, there is no use of that.

The current changes address this. In attendance, the start
and stop dates are now set to today instead of having the
start date as today and the stop date 2 days after.

In payroll, the start date is now set to the start of the current
month and the end date is now set to the end of the current month
instead of having the start date as the start of the current
month and the end date as the end of the month that comes
after the current month by 2.

targets: 17.3
task-3950880

closes odoo/enterprise#63412

Signed-off-by: Jurgen Gjini (jugj) <jugj@odoo.com>