[FIX] documents: no traceback when lost access to workspace multicompany
Issue: When we have multiple companies and we create different folders
for each other, if we are working in a folder from let's say company A
and then we switch to Company B, we will run the same link to this
folder but from the Company B rights point of view, and since this
folder is not accesible from this company it will throw a traceback and
break the page until refresh to another url.

Steps to reproduce:

1. Install Documents.
2. Have atleast 2 companies.
3. Create a new folder that belongs to the current company.
4. While having this folder selected, change to the 2nd company.

Solution:

The issue is triggered within the load funcion for the
DocumentsSearchModel, since we get the folderId that's in the url, and
we also get the list of the folder we have access to, we could filter
out if the current folder we are trying to access is within this list
before calling toggleCategoryValue, and If it's not we could just call
it on `folderId = false;` as default (which is the "All" section of
folders). This way we avoid calling `category.values.get(...)` later on
for undefined (we don't have access to this folder).

opw-3971935

closes odoo/enterprise#68459

X-original-commit: a219e9a8df4a7d88add364e2a5df9881a42680b1
Signed-off-by: Maruan Aguerdouh Mohtar (magm) <magm@odoo.com>