[FIX] l10n_es_reports: fix several issues in Libros Registro de IVA
Steps to reproduce:
- Install l10n_es_reports
- Switch to a Spanish company (e.g. ES Company)
- Create some invoices/bills and credit notes using different types of taxes
- Go to "Accounting / Reporting / Statement Reports / Tax Report"
- Select "Generic Tax Report"
- Click on "VAT Record Books (XLSX)" to download the report

Issues:
1) Credit notes appear as positive in the report
2) Withholding taxes (having type "Retencion") are not handled.
They should appear in the dedicated column.
3) Intra-European bills have a tax that is reverted. So the tax amount
should be 0, but it is not reverted in the report.
4) If an invoice has 2 lines that use 2 different taxes of type
"Recargo de Equivalencia", one of them will not appear in the report.
5) Advanced combinations of taxes on several invoice lines are not supported.

Solution:
This fix changes the way of generating the Libros Registro de IVA report.
Before this fix, the report was generated based on the tax lines.
One line of the report was corresponding to a regular tax line, which didn't
support advanced combinations of taxes on several invoice lines.
This fix analyzes the taxes repartition on the invoice lines and tries to
respect it when generating the report.
A line of the report corresponds to a combination of taxes of different type:
(IVA + surcharge tax + withholding tax)

Limitation:
Having 2 taxes of the same type on the same invoice line is not supported.
The total and the tax amount will be correct but not the tax rate because
a report line has only one column to display each type of tax.

opw-3844967

closes odoo/enterprise#64681

Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>