[FIX] documents: keep the related record on the new attachment
Steps to reproduce:
-------------------

- Install `Documents` and `Inventory` (for test purpose) modules
- Go to Settings and enable `Centralize files attached to products`
- Go to Inventory > Products > Choose any product
- Add an attachment to the Product
- Go to Documents, then select the `Products` workspace
- Select document related to product just uploaded
- `Replace` the attachment with a new one (old one should be added to
  `History`)

Issue:
------

The new attachment, and therefore the document, is not related to the
product anymore.

Cause:
------

When uploading a new attachment, we link the attachment to the document
and we don't keep the related record.

Solution:
---------

Keep the related record on the new attachment and remove it from the old
attachment to not have all versions of the attachments displayed on
the attachments of the products.

opw-3932497

closes odoo/enterprise#65512

X-original-commit: 0527b75de58e00fba2e417e5c0bb97d2a9de0591
Signed-off-by: Nasreddin Boulif (bon) <bon@odoo.com>