[IMP] sign: use the same date format when auto-filling date fields
Before this commit, it was possible to fill two date
fields by two different date formats in the same document.
This happend because the date format was
dependent on the location of user who is signing.

After this commit, date fields will be auto-filled by a fixed
date format, that format depends on the language of the company's partner.

closes odoo/enterprise#64591

Task: 3930358
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>