[FIX] stock_barcode: Prevent unnecessary line split on dest location change
# How to reproduce
- Create product: P1, storable | tracked by lot | barcode set
- Create Receipt for 2 units of P1
- In barcode App:
  - Scan product "P1"
    - Scan lot "lot1"
    - Scan lot "lot2"
  - Select line with "lot1"
  - Scan dest location "2601892" (Stock/Shelf1 in runbot)
  => New line created for P1 with quantity 0

The selected line is always done first.
Usually the selected line is the last one without reserved quantity, it is not split, and we steal the reserved quantity of the sibling.
However, if the line with incomplete reserved quantity is done first, then it is split. When the next line is done, it steals quantity from the created line, leaving it with reserved_uom_qty = 0 and qty_done = 0.

OPW-3942171

closes odoo/enterprise#65135

X-original-commit: 42e9c378b0766e98a93605d0c2bc0797e3995734
Signed-off-by: Steve Van Essche <svs@odoo.com>
Signed-off-by: David Fesquet (dafr) <dafr@odoo.com>