[FIX] account_reports: remove unnecessary action domain
A string domain is being used without being evaluated, leading to a
traceback

Steps to reproduce
------------------
* Turn on debug mode
* Go to the “Journal Entries” window action via Settings > Technical >
  Actions > Window Actions
* Give any domain to the “Domain Value” field of the window action
* Go to Tax Report via Accounting > Reporting > Statement Reports > Tax
  Report
* Click the “Closing Entry” button
* Click the “Post” button to post the closing entry

 you will run into the following traceback: `IndexError: string index
out of range`

Fix
---
The `account.action_move_journal_line` (Journal Entries window action)
record was being reused for consistency. However, domains added to that
record should not be propagated in this context. So in this case, the
solution is to remove the domain from the returned action.

opw-3903500

closes odoo/enterprise#64597

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>