[IMP] mrp: specify batch size in MPS to auto-split manufacturing orders
This commit adds a new field to `mrp.production.schedule` model:
`batch_size`. The field is shown and used only if the selected route is
a manufacturing route. If this value is specified, the quantity of the
generated manfacturing orders by MPS will be capped by this value,
hence, the MOs will be automatically split in this case.

Task-3962134

closes odoo/enterprise#64600

Related: odoo/odoo#169258
Related: odoo/upgrade#6186
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>