[FIX] account_reports: changes default value hide 0 filter
Issue:
It is possible to select the option "hide 0 lines" in the partner ledger.
For one, it does not make sense as if a partner appears, it means that there are transactions. Even if the balance is set to 0, we should display the transactions.
And then, the option simply does not work in partner ledger.

Solution:
Put as a default value "never" so it does not show up.
As it is the case in 16.0

opw-3981939

closes odoo/enterprise#64741

X-original-commit: 9514a3cd1752240bf1e78b2b0bacf81b3b74afd5
Signed-off-by: Olivier Colson (oco) <oco@odoo.com>
Signed-off-by: Yolann Sabaux (yosa) <yosa@odoo.com>