[FIX] documents_project: view all documents in project updates
Steps to reproduce:
- Install `documents_project`
- Go to one of the task and add a document
- Close it

Issues:
In the project update the number of documents is shown however it is
wrong, as it compute it by retrieving all the tasks without taking into
account the state of the tasks.

When clicking on the documents button you don't see documents linked to
task that are closed, the desired behaviour is to see all of them and
not take into account the state of the tasks.

The behaviour was due to the way the domain is set on the `task_ids` in
the project model.

https://github.com/odoo/odoo/blob/174cb193592f12c7f4167eb34642665778deb208/addons/project/models/project.py#L332-L333

Because of this when we retrieved the documents we didn't have all the
tasks.

https://github.com/odoo/enterprise/blob/0e98a5fdf5e4f32eeae769552284297b9233088c/documents_project/models/project_project.py#L204

opw-3921156

closes odoo/enterprise#64720

X-original-commit: 3ab8f567f95af926d88ba99432393dd851bde507
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Mattis Megevand (mmeg) <mmeg@odoo.com>