[IMP] l10n_br_avatax: inform Avatax of payment terms
To conform to government and customer expectations, we have to inform
them of the payment terms set on the invoice.

The vNet, vOrig, grossValue and netValue keys are required but meant
to record "payment discounts". In Odoo the closest concept is the
early_discount feature on payment terms. But we don't know if this
early discount applies when we e-invoice, it's only known when the
customer pays. Because of this, we always send the full amounts in
these fields. At a later stage we can see if/how we should support
this.

task-3968325

closes odoo/enterprise#64953

Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>