[IMP] l10n_be_hr_payroll: use exoneration in termination fees withholding taxes computation.
A CP200 employee may be eligible for exoneration
based on the number of dependent children they have.
This change takes the value of exoneration into
account in the computation of termination fees withholding
taxes. It also refactors the computation of withholding taxes
for double holiday pay, 13th month and termination fees into
one function.

task-3909012

closes odoo/enterprise#74019

X-original-commit: 418d95295fcd4216f174522cc9f67331dd71106f
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>