[IMP] l10n_be_hr_payroll: use exoneration in termination fees withholding taxes computation.
A CP200 employee may be eligible for exoneration
based on the number of dependent children they have.
This change takes the value of exoneration into
account in the computation of termination fees withholding
taxes. It also refactors the computation of withholding taxes
for double holiday pay, 13th month and termination fees into
one function.

task-3909012

closes odoo/enterprise#73665

X-original-commit: 4066e438062ab0f7be24ffd34f5cd6bfee5de06e
Signed-off-by: Bertrand Dossogne (bedo) <bedo@odoo.com>