[FIX] documents: fix move locked document
Before this commit, when dragging documents including one or more locked
documents to another workspace, only the unlocked documents are indeed moved
and no message is displayed to alert the user although the code exists but was
never triggered.
This commit fixes this issue by displaying a notification informing the user
that the locked files haven't been moved.

Task-3992114

X-original-commit: bbe77c5ce3dfeea9fb7fc695f65f96c2f39f033d
Part-of: odoo/enterprise#65723
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>