[FIX] documents: fix move locked document
Before this commit, when dragging documents including one or more locked
documents to another workspace, only the unlocked documents are indeed moved
and no message is displayed to alert the user although the code exists but was
never triggered.
This commit fixes this issue by displaying a notification informing the user
that the locked files haven't been moved.

Task-3992114

X-original-commit: e94ed19e914a657db34fa5a642d463b2cabb4bf1
Part-of: odoo/enterprise#66695
Signed-off-by: Warnon Aurélien (awa) <awa@odoo.com>