[IMP] sign: load only one font by default
Before this commit, when the user opened the sign dialog, all the fonts
were downloaded, which could be slow depending on the internet
connection.

We decide to use the lightest default font for automatic signature.
Other fonts are downloaded when the user goes to the 'Auto' tab. They
are downloaded in the background.

Taskid: 3713718
Part-of: odoo/enterprise#64685
Related: odoo/odoo#169507
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>