[IMP] project: {re,}schedule multi users tasks
Before this commit:
- {auto planning / moving forward and backward} tasks with multi
users is following the company calendar leading to create conflicts.
- a bug was found, the first possible date to plan is taking into
account the deadlines of the tasks that will be planned. for example
if task A is blocking task B and the deadline of A is 01/05/2023,
first planned date to plan for B will be 01/05/2023
then if A was planned with a new deadline before the old one 01/04/2023
so the old deadline should no more impact task B planned_date_begin.
- It's not possible to undo the scheduling if the user is not satisfied
with the result.

After this commit:
- tasks are planned in a time that would be suitable for all users.
- old deadlines of tasks to plan are no more taken into account.
- An 'undo' button in a notification toast after a smart
scheduling in the project task gantt view has been triggered. This
allows the user to rollback to the previous state.

Technically,
for move back/for{ward}, kahn algorithm with priority queue was changes
to a simple for loop.

why ? before this commit, an index is stored and keep incrementing for
each interval. There is no need to iterate intervals again as we're
pretty sure that previous intervals are used and the coming ones are
not used

but in this commit, a user availabilities can be in his intervals and
in another interval for his availabilities and another user, so there is
no way to be sure that coming intervals are not used as they can be used
in another context.

keep in mind, that in case of perf problems, it's possible to implement
again kahn approche and verifying if the coming interval is good or not
by comparing it with the used for planning, it's a bit complicated to
implement and maintain. but it's still a possible optimized solution if needed.

task-3853065

closes odoo/enterprise#64206

Related: odoo/odoo#174887
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>