[ADD] l10n_br_avatax_sale: add support for CNAE codes
A CNAE code determines the type of activity a company is involved
in. Companies can be registered to have one or more
codes. Transactions are linked to these codes, depending on the exact
nature of them.

To support it in Odoo we add a default, primary CNAE code on the
company. This is used as a default on both sale.order and
account.move. In case a transaction is linked to a secondary activity
sector of the company, the user can change it.

The _l10n_br_is_avatax() function in the mixin was changed to a
non-stored computed field so it can be used to determine visibility of
the new fields in the views.

The res_config_settings_views.xml file was re-formatted because its
indentation was inconsistent.

task-3987919

Part-of: odoo/enterprise#65942
Related: odoo/odoo#171856
Signed-off-by: Josse Colpaert (jco) <jco@odoo.com>