[FIX] l10n_co_dian: display supp doc btn when journal allows it
The "Send Support Document to DIAN" button was displayed on vendor bills
even when the associated journal was not set up for it (the field
"Resolución de Facturación" was not set up for that move's journal).
The button is now invisible if the correct conditions are not met.

task-3990920

closes odoo/enterprise#64917

X-original-commit: ee19ad54a50534c974412a74da180e311a7af837
Signed-off-by: Julien Van Roy (juvr) <juvr@odoo.com>