[IMP] web_studio: allow pushing a state key that won't be part of the URL
This commit introduces the possibility to add a key/value to the state,
but omit it on the URL, a new function `addKeystoOmit` was added to this
effect. Note that, this key will also be added in the state that is sent
to the browser history. Allowing to retrieve the complete state
(including the omitted keys) when clicking the back button on the
browser.

X-original-commit: f3e979995b5e327073c2e2b788667fb1dd670558
Part-of: odoo/enterprise#65468
Related: odoo/odoo#170792
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>