[FIX] web_enterprise: list button for upselling is registered just after its parent
web_enterprise modifies the js template for the list_renderer by introducing an upsell point for Studio
in the optional columns dropdown.

That inheritance spec was not at any particular place in the inheritance hierarchy.
On some views (e.g. account invoices) this feature was not available.

This commit fixes the issue by manually forcing the extension in web_enterprise to go
just after list_renderer.xml from the web module.

After this commit, the upsell feature is available on the views that did not have it before.

closes odoo/enterprise#65357

X-original-commit: 622202bb62bab304428cb359d454c57596b67707
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>