[IMP] web, *: review z-index utilities
This commit removes the custom `z-index` utilities in favor of Bootstrap
default to provide more flexibility.

Prior to this commit, we were using custom utilities to handle
the z-index of our elements. This approach allowed us to use `z-index-0`
and `z-index-1` classes but it was sometimes not enough,
leading to the use of `z-index` CSS properties.

With the migration to Bootstrap 5.3 achieved in Commit[1],
we now have access to a wider range of `z-index` which are defined inside
the `$zindex-levels` map and go from `-1` to `3`. Unfortunately,
these new utilities were being overridden by our custom one.

To unlock these utilities and allow more flexibility in our code, this
commit removes the custom utilities in favor of Bootstrap default.

Commit[1]: 6e90b00

task-3930430

closes odoo/enterprise#62524

Related: odoo/odoo#165568
Related: odoo/upgrade#6078
Signed-off-by: Pierre Paridans (app) <app@odoo.com>