[IMP] account_reports: allow dict in markup for line ids
Improves slightly the markup of account report lines ids such that it can
contain a dictionary, therefore enabling a key-value pair. Previously,
it was made using strings, which was quite a hacky way of doing. This
markup is used for example either to group lines, indicate a line is
a total line, or for 'load more' lines. We standardize this by putting
json into the markups in case of dictionary.

task-3791247

closes odoo/enterprise#65087

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>