[FIX] l10n_be_reports: removing empty xml tag comment on tax report
When exporting tax report to xml, a comment tag is added in the xml even when empty or set to False.
Or per https://finances.belgium.be/fr/E-services/Intervat/comment-utiliser-intervat/deposer-declaration-periodique#q5
tags cannot be empty

Steps to reproduce:
1.In a belgian company > accounting > Reporting > Tax Report
2.Export in xml
3.Do not fill information in the pop up > Export XML
4.The <comment> tag displays False when no comment

cause:
comment is set to False when empty, thus display False.

Solution:
set comment to '/' when empty

opw-3980954

closes odoo/enterprise#64798

Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>