[FIX] stock_barcode_mrp: not redirect to backend from barcode backorder
Steps to reproduce:

- Enable multi-step routes in the settings
- Inventory > Configuration > Warehouse Management > Operations types
- Click on Manufacturing and put "Always" on create a backorder
- Create a product P (no need for a bom)
- Create and confirm a manufacturing order for 2 units of the product
- Go to the barcode module > operations > manufacturing > the MO
- Register only one unit of P and produce

> A backorder is automatically created but you are redirected to the
  backend.

Note: If you were in "ask" on create backorder or if your product had a
bom and you did not register the quantities of consumed components, a
pop up would appear and resolving the pop up would redirect you to the
kanban view of mrp.production in the barcode module.

Cause of the issue:

The `validate` method called by clicking will execute the action if the
orm call returns one and will add an 'onClose' call to its options:
https://github.com/odoo/enterprise/blob/661fd61c0f50d4bf80484227c92a3fd0b9a05c68/stock_barcode/static/src/models/barcode_model.js#L475-L482
https://github.com/odoo/enterprise/blob/661fd61c0f50d4bf80484227c92a3fd0b9a05c68/stock_barcode/static/src/models/barcode_model.js#L487-L492
If the action is a popup, the onClose method will be called once the
pop up is resolved. However, if the action redirects you dirrectly
to the form view of the mrp.propduction the onClose method will not be
triggered and you will not be redirected to the kanban view.

Fix:

We change the return of the action_backorder in case it is triggered
from the barcode module.

Follow up of commit 6915741

opw-3890886

closes odoo/enterprise#65279

X-original-commit: d90acab45ca8ba9357c3b3e85c396c0323ae28d6
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Lancelot Semal (lase) <lase@odoo.com>