[FIX] mrp_plm: fix kanban ghost height
This commit aims to fix an issue about the `kanban_ghost` elements within
PLM having a visible white background.

Commit[1] introduced a simplified version of our kanban archs. Within
this commit, a white background is applied on
`.o_kanban_record:not(.o_legacy_kanban_record)`, meaning this also
applies to `o_kanban_ghost` elements.

To avoid this, a `max-height` was set to `O` on `o_kanban_ghost` elements
to prevent them from being visible.

While this handle most of the scenario, it appears that
there is a specific rule within the `mrp_plm` module that sets
a `min-height` to `o_kanban_record`, making them visible due to the white
background.

To handle this issue, we set the `min-height` property only to other
elements that are not `o_kanban_ghost`.

Commit[1]: 8995473

task-3996625

closes odoo/enterprise#64851

Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>