[FIX] planning_contract: get contract end date
Steps to reproduce:
-------------------
With a user which has no access rights for 'Employees' and
'Contracts' and with admin rights for 'Planning', tries to
add a new recurring shift (via the Planning module).

Issue:
------
We get an access error.

Cause:
------
We're trying to access the `contract_ids` field even though
we don't have access rights.

Solution:
---------
Fetch the value of the `date_end` field in sudo so
that this flow is not blocked for a planning user.

opw-3992135

closes odoo/enterprise#65522

X-original-commit: c9a8d8f1e3439c1adc1f221dc92c549d62534c5d
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Thomas Lefebvre (thle) <thle@odoo.com>