[FIX] iot: websocket always listen to first company channel
Before this commit:
An IoT box would always listen to the iot.channel of the first
company as it did not take into account the IoT-box company field.
As such, setting an IoT box on a company C2 setting a printer and
report to print on this IoT and restarting it.
Report will never print as the IoT listen to C1 channel

After this commit:
If the IoT box company is set and the IoT is restarted. It will
received the correct IoT channel. So, to print IoT reports

opw-3991810

closes odoo/enterprise#65008

X-original-commit: 86eba882a800a8d1e16608783ef419c43d03f1ce
Signed-off-by: Loan Sens (lse) <lse@odoo.com>
Signed-off-by: Yaroslav Soroko (yaso) <yaso@odoo.com>