[FIX] helpdesk: update internal user phone on ticket
Steps to reproduce:
- Install Helpdesk
- Create two internal user with admin access to Helpdesk
- Log as the first user and create a ticket for the second user with
  empty phone

Issues:
An access error is shown as we do not have the rights to write on the
partner phone. To fix this issue a sudo was added as well as a filter in
the inverse to make sure we skip ticket without a partner phone in the
inverse.

opw-3989614

closes odoo/enterprise#65402

X-original-commit: 7c83ea1c3103abe790519b7bb3229b2c02d77570
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>