[IMP] sign: improve linked record integration
Since
https://github.com/odoo/enterprise/commit/5f347b7756394262a350d6c2cbbb0c4d780a4d58
,it is possible to link a record to a sign.request
but it was only handled when the user created an activity of type signature.

After this commit, when a sign.request is created from another app, the
linked record will allow to go back to it's form view when the signature
is done as a loggedin user.

Taskid: 3721057
Part-of: odoo/enterprise#65010
Related: odoo/upgrade#6206
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>