[FIX] voip: fix test failure
"Switch audio input" test failed in an undeterminitic manner on this
line:
https://github.com/odoo/enterprise/blob/e808e583e9d406735a05dab42bf24f8b89899b67/voip/static/tests/softphone/device_selection_dialog.test.js#L54
After clicking the "Call" button, func `makeCall` has asynchronous call
before setTimeout:
https://github.com/odoo/enterprise/blob/e808e583e9d406735a05dab42bf24f8b89899b67/voip/static/src/core/user_agent_service.js#L268
https://github.com/odoo/enterprise/blob/e808e583e9d406735a05dab42bf24f8b89899b67/voip/static/src/core/user_agent_service.js#L284
It may happen that because of the asynchronous call, our
`advanceTime(5000)` is executed before we call `setTimeout` in `makeCall`.
To avoid this, we check the dom change to make sure the asynchronous call
has finished before we continue.

closes odoo/enterprise#65246

X-original-commit: 7065a4500e5351cc75323b2efc4a85ea0cf3a0e0
Signed-off-by: Louis Wicket (wil) <wil@odoo.com>