[FIX] account_accountant: allow reconcilition with other liquidity accounts
After [this commit], when we allowed credit card accounts on bank
journals, we changed the domain of the manual bank reconciliation
widget to also make credit card accounts not selectable anymore.

However, people are currently reconciling their bank journal entries
with a credit card account, which is not possible anymore after this
change.

This commit fixes that by only preventing to reconcile with the same
account as the one set on the journal or an `asset_cash` or
`off_balance` account.

[this commit]: https://github.com/odoo/enterprise/commit/4f84a260e515cc5d52d7aa2d9ba0dca7d77a101c

closes odoo/enterprise#65176

X-original-commit: 31b081fe4f88ee8ded3ce147adc9636a4e3fea75
Signed-off-by: Laurent Smet (las) <las@odoo.com>