[IMP] sale_amazon: do not create new states
Before this commit contact created from amazon order create
new state if it does not find matching state in odoo and
it was causing issues like having record multiple record
for same state and user don't know which one is right with
proper code in data.

This commit stop creating new state instead create activity
on contact for amazon sale person to set proper state on
the related contact to not break flows that are related to
contact's state.

task-3865454

closes odoo/enterprise#64893

Signed-off-by: Kartik Chavda (kcv) <kcv@odoo.com>