[REF] *spreadsheet*: normalize values
This commit is the counterpart of https://github.com/odoo/odoo/pull/169914.
It adapts the autofill plugin and its tests to the parsing of arguments.

It also adapts the tests to the new formatting of arguments (boolean is
now a boolean in the formula, a number is a number, etc.).

Task: 3991743
Part-of: odoo/enterprise#65379
Related: odoo/odoo#170661
Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>