[FIX] partner_commission: check template if necessary
Steps to reproduce:
-------------------
- create a commission plan with:
    product category: "All"
    rate: R
    capped: True
    max commission: >>>
- create a partner level with the created commission plan
- create a partner with the created partner level

- create a product in the correct category with price P
- create a quotation template with the created product

- create a sale order with the partner as referrer
- add the quotation template (the product is added automatically)
- confirm
- create the invoice, confirm it and register the payment

At this moment, a purchase order is created (the commission)
If we repeat these steps with a new sale order,
the amount of the purchase order will increase as expected.

- add a quotation template on the commission plan (in the rules)
- repeat the creation of a sale order.

No commission is added (purchase order line).

Cause:
------
We have no template when calling `_match_rules`.
However, our plan has a default template.
The result is that no rules are found.

Solution:
---------
Fallback on the sale order's template
if it is not a subscription.

opw-3933323

closes odoo/enterprise#64884

Signed-off-by: Pierre Masereel (pim) <pim@odoo.com>