[FIX] web_gantt: reinitialize column mappings earlier
The reinitialization of mappingColToColumn and mappingColToSubColumn
was done too lately in computeDerivedParams. This used to invalidate
some computations done in functions called by computeDerivedParams like
addTo in MRPWorkorderGanttRenderer (we could get a pillDuration of let's
say -3456 hours for example).

Here we make sure that the mappings are reinitialized at the beginning
of computeDerivedParams.

closes odoo/enterprise#64918

X-original-commit: c28f80f476612f290074640f21aee5e7a0c72839
Signed-off-by: Bruno Boi (boi) <boi@odoo.com>
Signed-off-by: Mathieu Duckerts-Antoine (dam) <dam@odoo.com>