[FIX] account_bank_statement_import_camt: secure the way we get transaction types
The issue:
Our code generally assumes that strings are uppercase, which is typically true. However, in some cases, such as the one described in this ticket, the input string is only capitalized (the first character in uppercase).

The Fix:
convert the keys to uppercase before retrieving their values from the array, and use the get method instead of the bracket notation for safer access.

opw-3979945

closes odoo/enterprise#65576

X-original-commit: 33ff107b525f7190507ad453ab43f3cf27927ecc
Signed-off-by: Claire Bretton (clbr) <clbr@odoo.com>