[FIX] sale_subscription : Show optional product when have invoice start date
Steps to reproduce:
	- Install Sale Subscription module
	- Create a recurrent order and add Optional products
	and Start date under Other info tab
	- Click on Preview

Current behavior before PR:
You won't see the optional products in the portal view if for a recurrent
order that has start date and in draft. This is happening because in
sale_subscription we are checking if the start_date of the order is equal to
the next_invoice_date
https://github.com/odoo/enterprise/blob/17.0/sale_subscription/models/sale_order.py#L1957:1959
and if the order is in draft it won't have a next_invoice date until it is
either confirmed or sent.

Desired behavior after PR is merged:
We are now checking if the next_inovice_date is not set we show the
optional products.

opw-3980405

closes odoo/enterprise#66285

X-original-commit: e667f0096aac58f9bc66b7f0066cc57b35962079
Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>
Signed-off-by: Youssef Bashandy (yoba) <yoba@odoo.com>