[FIX] data_merge: remove messages linked to unlinked records
Steps to reproduce:
-------------------

- Install `data_merge` and `CRM` module
- Go to CRM and create 2 opportunities with same name
- Activate debug mode and go to `Scheduled Actions`
- Run `Data Merge: Find Duplicate Records` action
- Delete the Deduplication Rule for `Lead/Opportunity`
  (or uninstall `CRM` module)
- Go to Discuss and click on `History` channel

Issue:
------

Error: `Record does not exist or has been deleted.`.

Cause:
------

When uninstalling a module that have a model that is used as `res_model_id`
in a `Deduplication Model`, we do not remove message the message linked
to the record deleted.
Therefore, when opening history channel, we try to access messages
linked to records that might have been deleted.

Solution:
---------

When removing `Deduplication Model`, remove also the messages linked to
the record.

opw-3906945

closes odoo/enterprise#67094

X-original-commit: 2aa518fb68fdcee2e5893451618279114ad0bc40
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>