[FIX] website_sale_renting: retain rental period on attribute selection
-16.0

Steps:
-Install rental app.
-Activate the date picker from the web-editor.
-Add the start date and end date from that.
-Select some attribute of the product.

Issue:
-When any user adds the start date and end date from the date picker
and after that, the user adds the other attribute(filter) of the product, that
selected rental period gets cleared automatically.

Cause:
-The selected rental period (start and end dates) is not persisted when
the page refreshes or updates due to attribute selection.
-The input fields for the rental period are not updated properly,
causing the dates to be reset.

Fix:
Extend the '_onChangeAttribute' method to read the start_date and end_date
parameters from the URL.
Update the hidden input fields for 'start_date' and 'end_date' with the
values before applying the attribute changes.

opw-3774060

closes odoo/enterprise#67562

X-original-commit: 74dda283cbd818b819ad46f9296ca0e0582803d2
Related: odoo/odoo#174904
Signed-off-by: Valentin Chevalier <vcr@odoo.com>