[FIX] l10n_be_reports: replace `user_has_groups` by `has_group`
In 17.0, a protection was added to the 281.50 tags on the partners
(commit: 09439b46cbb8b6ef2447d67e1647cbc284d0e192).

The code was forward ported but the function `user_has_groups` has been
removed in saas-17.2 (in
https://github.com/odoo/odoo/commit/cb31d4352c741a5f8a927a65debb8e027e87df22),
causing a traceback. In this commit, we rather use the `has_group`
function.

task-3955109

closes odoo/enterprise#65033

X-original-commit: cb97a830c19b1735da718efc62f9bf88610c72a3
Signed-off-by: Florian Gilbert (flg) <flg@odoo.com>