[IMP] knowledge: fix error 403 on homepage for public users
When the public users access the `/knowledge/home` route, the system will
open the first accessible, visible and published article using a domain
involving different search compute methods.

In the `_search_is_article_visible` search method, the system will fetch
the members associated to the current user. However, the public users do
not have access to that model. As a result, the public user will get an
access error and the server will display the 403 error page.

Steps to reproduce the error:
1. Install `website_helpdesk_knowledge` module.
2. On the Odoo homepage, click on the helpdesk icon.
3. On the helpdesk overview, click on the vertical dots of the "Customer
   Care" team card.
4. On the dropdown menu, click on the "Settings" option.
5. On the "Help Center" section, click on the "Knowledge" option.
6. On the "Visibility & Assignment" section, click on "Invited portal
   users and all internal users" option.
7. Save the changes.
8. Open a new browser in incognito and go on your website.
9. Click on the "Customer Care" navbar link of your website.
10. Click on the "Browse articles" button
11. You should get a 403 error page
=> TO BE: You shouldn't get a 403 error page.

Since the public users cannot be added to the article members list, we
will skip the members fetching for those users and return an empty domain.
The condition will then be ignored and the search method will not raise
any exception.

task-3980714

X-original-commit: e7b8be376eb3b94ef1cb2b6cecda891c4227c52e
Part-of: odoo/enterprise#68360
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>