[IMP] l10n_mx_edi_stock{_extended}_{30,31}: add Carta Porte 3.0 PDF
[This commit] added in a new Carta Porte 3.0 PDF for Mexico in version
saas-17.3 that inherits from the standard delivery slip, by removing
all the tables related to the products etc. (for security reasons: the
driver doesn't need to know what's in the cargo) and adding in some new
required fields.

This commit backports the report from saas-17.3 with the following
changes:
- We improved the layout a bit to make the report even clearer.
- We corrected the "Num Reg ID Trib" and "Residencia Fiscal" of the
  "Figura"'s to use the right value and show them per figura.

We also fixed a bug introduced by [this fw-port] to saas-17.2. The added
`RegimenesAduaneros` were still using the namespace `cartaporte20`,
while everything is saas-17.2 was already changed to `cartaporte30`.
Hence the namespace did not get updated to `cartaporte31` and the
validation of the CFDI failed.

These changes will be forward-ported again to the upper versions.

[This commit]: https://github.com/odoo/enterprise/commit/af14612ec5942acdc06584ec8ff9d8bf26998ae7
[this fw-port]: https://github.com/odoo/enterprise/commit/659f491423ee0680b103bd4dfe483dae67deb03e

task-3990042

closes odoo/enterprise#69970

X-original-commit: ba1a34f9a2252b10de194b31da1613b00642360c
Signed-off-by: John Laterre (jol) <jol@odoo.com>
Signed-off-by: Dylan Kiss (dyki) <dyki@odoo.com>